moved did events, pruned did-spi dependencies
