+++ e2e-tests/api-tests/src/test/java/org/eclipse/edc/identityhub/tests/DidManagementApiEndToEndTest.java
@@ -16,8 +16,8 @@ package org.eclipse.edc.identityhub.tests;
 
 import io.restassured.http.Header;
 import org.eclipse.edc.iam.did.spi.document.DidDocument;
-import org.eclipse.edc.identityhub.spi.events.diddocument.DidDocumentPublished;
-import org.eclipse.edc.identityhub.spi.events.diddocument.DidDocumentUnpublished;
+import org.eclipse.edc.identithub.spi.did.events.DidDocumentPublished;
+import org.eclipse.edc.identithub.spi.did.events.DidDocumentUnpublished;
 import org.eclipse.edc.identityhub.spi.participantcontext.model.ParticipantContext;
 import org.eclipse.edc.junit.annotations.EndToEndTest;
 import org.eclipse.edc.spi.event.EventRouter;
