+++ extensions/api/management-api/did-api/src/main/java/org/eclipse/edc/identityhub/api/didmanagement/DidManagementApiExtension.java
@@ -14,8 +14,8 @@
 
 package org.eclipse.edc.identityhub.api.didmanagement;
 
-import org.eclipse.edc.identithub.did.spi.DidDocumentService;
-import org.eclipse.edc.identithub.did.spi.model.DidResource;
+import org.eclipse.edc.identithub.spi.did.DidDocumentService;
+import org.eclipse.edc.identithub.spi.did.model.DidResource;
 import org.eclipse.edc.identityhub.api.didmanagement.v1.DidManagementApiController;
 import org.eclipse.edc.identityhub.api.didmanagement.v1.GetAllDidsApiController;
 import org.eclipse.edc.identityhub.spi.AuthorizationService;
